✨ update  : 피해감소 -> 방어로 변경.

+++ TextRpgProject/Character.h
@@ -36,7 +36,7 @@ private:
 	int experienceCapacity;				// 경험치 통
 	int gold;							// 골드
 
-	double damageReduction;				// 피해감소 STR
+	int damageReduction;				// 피해감소 STR
 	double accuracy;					// 명중률 DEX
 	double skillEnhancement;			// 스킬강화 INT
 	double criticalChance;				// 치명타확률 LUK
@@ -69,7 +69,7 @@ public:
 	int GetExperience() const;
 	int GetExperienceCapacity() const;
 	int GetGold() const;
-	double GetDamageReduction() const;
+	int GetDamageReduction() const;
 	double GetAccuracy() const;
 	double GetSkillEnhancement() const;
 	double GetCriticalChance() const;
@@ -87,7 +87,7 @@ public:
 	void SetExperience(int _experience);
 	void SetExperienceCapacity(int _expCapacity);
 	void SetGold(int _gold);
-	void SetDamageReduction(double _damageReduction);
+	void SetDamageReduction(int _damageReduction);
 	void SetAccuracy(double _accuracy);
 	void SetSkillEnhancement(double _skillEnhancement);
 	void SetCriticalChance(double _criticalChance);
